Moved cli args to new file with validation
