+++ cli/src/lib.rs
@@ -1,35 +1,14 @@
+use clap::Parser as ClapParser;
+use cli_args::CliArgs;
+use code_generation::CodeGenerator;
+use parser::Parser;
 use std::{
     error::Error,
     fs,
     io::{self, Write, stderr, stdout},
-    path::PathBuf,
 };
-
-use clap::Parser as ClapParser;
-use code_generation::CodeGenerator;
-use parser::Parser;
 use transpiler::Transpiler;
-
-#[derive(ClapParser, Debug, PartialEq)]
-pub struct CliArgs {
-    #[arg(short, long, value_parser = file_extension_validation)]
-    input_path: PathBuf,
-
-    #[arg(short, long, default_value = "./from.js")]
-    output_path: PathBuf,
-}
-
-pub fn file_extension_validation(input_path: &str) -> Result<PathBuf, String> {
-    let path_buf = PathBuf::from(input_path);
-
-    match path_buf
-        .extension()
-        .and_then(|input_path| input_path.to_str())
-    {
-        Some("from") => Ok(path_buf),
-        _ => Err(format!("Could not find or process {}", path_buf.display())),
-    }
-}
+mod cli_args;
 
 #[derive(ClapParser, Debug)]
 pub struct Cli;
@@ -65,6 +44,7 @@ impl Cli {
 #[cfg(test)]
 mod cli_tests {
     use super::*;
+    use std::path::PathBuf;
 
     #[test]
     fn cli_args() {
